refactor(web): extract AppLayout shell (sidebar + theme/lang header)
Pulls the chrome out of DashboardView into a reusable AppLayout: sidebar
rail + a header with a #title slot (left) and the global theme/language
controls (right), then a default slot for the body. NodeDetailView (next
commit) reuses it, so the header isn't duplicated — the shared-layout
concern from the Stage 2a review.

- DashboardView now wraps its content in <AppLayout> and supplies its
  title via the slot; data-test="app-shell" lives on AppLayout's root so
  Playwright waitForAppShell stays valid.
- Theme/language coverage moves to AppLayout.spec; DashboardView.spec
  keeps the dashboard-body structure assertions.
