+++ nodelite-server/web/src/components/AppLayout.vue
@@ -1,0 +1,143 @@
+<script setup lang="ts">
+import SidebarNav from '@/components/SidebarNav.vue';
+import { useTheme } from '@/composables/useTheme';
+import { useLanguage } from '@/i18n/language';
+import { SUPPORTED_LOCALES, type SupportedLocale } from '@/i18n';
+
+// Shared chrome for every top-level view: sidebar rail + a header whose
+// left side is a per-view #title slot and right side holds the global
+// theme + language controls. Body goes in the default slot. Lives in one
+// place so DashboardView / NodeDetailView don't each reimplement it.
+const { theme, toggleTheme } = useTheme();
+const { currentLocale, setLocale } = useLanguage();
+
+function onLanguageChange(event: Event): void {
+  setLocale((event.target as HTMLSelectElement).value);
+}
+
+const localeLabels: Record<SupportedLocale, string> = {
+  en: 'English',
+  'zh-CN': '中文',
+};
+</script>
+
+<template>
+  <div class="app" data-test="app-shell">
+    <SidebarNav />
+
+    <main class="main">
+      <header class="page-header">
+        <div class="page-title">
+          <slot name="title" />
+        </div>
+        <div class="page-actions">
+          <select
+            class="lang-select"
+            :aria-label="$t('common.language')"
+            data-test="language-select"
+            :value="currentLocale"
+            @change="onLanguageChange"
+          >
+            <option v-for="locale in SUPPORTED_LOCALES" :key="locale" :value="locale">
+              {{ localeLabels[locale] }}
+            </option>
+          </select>
+          <button
+            type="button"
+            class="theme-toggle"
+            :title="$t('common.theme_toggle')"
+            :aria-label="$t('common.theme_toggle')"
+            data-test="theme-toggle"
+            @click="toggleTheme"
+          >
+            <svg
+              v-if="theme === 'dark'"
+              class="sun"
+              viewBox="0 0 24 24"
+              fill="none"
+              stroke="currentColor"
+              stroke-width="2"
+              stroke-linecap="round"
+              stroke-linejoin="round"
+            >
+              <circle cx="12" cy="12" r="4" />
+              <path
+                d="M12 2v2M12 20v2M4.93 4.93l1.41 1.41M17.66 17.66l1.41 1.41M2 12h2M20 12h2M4.93 19.07l1.41-1.41M17.66 6.34l1.41-1.41"
+              />
+            </svg>
+            <svg
+              v-else
+              class="moon"
+              viewBox="0 0 24 24"
+              fill="none"
+              stroke="currentColor"
+              stroke-width="2"
+              stroke-linecap="round"
+              stroke-linejoin="round"
+            >
+              <path d="M21 12.79A9 9 0 1 1 11.21 3 7 7 0 0 0 21 12.79z" />
+            </svg>
+          </button>
+        </div>
+      </header>
+
+      <slot />
+    </main>
+  </div>
+</template>
+
+<style scoped>
+.app {
+  display: grid;
+  grid-template-columns: 72px minmax(0, 1fr);
+  min-height: 100vh;
+  background: var(--bg-app);
+  color: var(--text-primary);
+}
+.main {
+  padding: 24px clamp(20px, 3vw, 36px) 40px;
+  max-width: 1680px;
+  width: 100%;
+}
+.page-header {
+  display: flex;
+  justify-content: space-between;
+  align-items: flex-start;
+  gap: 24px;
+  margin-bottom: 22px;
+}
+.page-actions {
+  display: flex;
+  align-items: center;
+  gap: 12px;
+}
+.lang-select {
+  background: var(--bg-card);
+  color: var(--text-secondary);
+  border: 1px solid var(--border-soft);
+  border-radius: 10px;
+  padding: 6px 10px;
+  font-size: 12px;
+}
+.theme-toggle {
+  width: 36px;
+  height: 36px;
+  border-radius: 12px;
+  border: 1px solid var(--border-soft);
+  background: var(--bg-card);
+  color: var(--text-secondary);
+  display: grid;
+  place-items: center;
+  transition:
+    background 150ms ease,
+    color 150ms ease;
+}
+.theme-toggle:hover {
+  color: var(--text-primary);
+  background: var(--bg-card-soft);
+}
+.theme-toggle svg {
+  width: 18px;
+  height: 18px;
+}
+</style>
